FIX: data.insert() converts empty ?args= to None. #300
Earlier, .replace('', None) was leaving them as empty strings.

This commit also fixes several issues with latest Pandas.

Tested on:
pip install pandas==1.3.5 numpy==1.21.5
conda install -c conda-forge rpy2==3.4.5

+++ tests/utils.py
@@ -510,7 +510,7 @@ def test_function(
         i1: Annotated[int, 'First value'] = 0,  # noqa
         i2: Annotated[int, 'Second value'] = 0,  # noqa
         s1: str = 'Total',
-        n1: np.int = 0,
+        n1: int = 0,
         n2: np.int64 = 0,
         h: Header = '',
         code: int = OK):
